+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This changelog summarizes major changes between GraalVM versions of the Graal JavaScript (ECMAScript) language runtime.
 The main focus is on user-observable behavior of the engine.
 
+## Version 1.0.0 RC14
+* Option `js.experimental-array-prototype` has been renamed to `js.experimental-foreign-object-prototype`.
+  In addition to setting the prototype of array-like non-JS objects to `Array.prototype`
+  it sets the prototype of executable non-JS objects to `Function.prototype` and
+  the prototype of all other non-JS objects to `Object.prototype`.
+
 ## Version 1.0.0 RC13
 * Made Java interop available in native images. Note that you have to configure the accessible classes and methods at native image build time (see [reflection configuration](https://github.com/oracle/graal/blob/master/substratevm/REFLECTION.md#manual-configuration)).
 * Removed deprecated experimental `Java.Worker` API. Node.js Workers should be used instead.